requirements.txt
semantic-kernel
pydantic
python-dotenv
httpx
pypdf
python-docx
jsonschema
openai
